Changed cast
Doesn't seem important

+++ src/cryptonote_core/blockchain_storage.cpp
@@ -344,7 +344,7 @@ difficulty_type blockchain_storage::get_difficulty_for_next_block() {
   CRITICAL_REGION_LOCAL(m_blockchain_lock);
   std::vector<uint64_t> timestamps;
   std::vector<difficulty_type> commulative_difficulties;
-  size_t offset = m_blocks.size() - std::min(m_blocks.size(), static_cast<size_t>(DIFFICULTY_BLOCKS_COUNT));
+  size_t offset = m_blocks.size() - std::min(m_blocks.size(), static_cast<uint64_t>(DIFFICULTY_BLOCKS_COUNT));
   if (offset == 0) {
     ++offset;
   }
